🏗️🚧 Splited the privateConv's service into a file named privateConv.
